backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
tensorflow==2.20.0
numpy==1.26.2
Pillow==10.0.1
opencv-python==4.8.0.76
gunicorn==21.2.0
